Fix issue "About command crashes".
The isse is that from 2.0 days we used to show Studio version in About box. Currently there is no feture representing studio...so the build is not able to initialize the value of %studio.version% and that chokes Eclipse's about box messages formatting. For now we will not show the Studio version.