Implement "Add Outline view to Window->Show View".
When the Outline is shown it appears under the App Explorer.

Also make sure that it Console and other view that stack with it go in the bottom right area instead of bottom (full width of the window).

We have to use the programmatic mechanism to place the Terminal view to deal with the fact that Terminal Views can have multiple instances with different secondary ids.