Fix issue "The Automatic update has been forced should be remembered outside of the workspace - possibly in configuration scope"
The fact that Automatic Update Check was forced is recorded in the Instance scope. Thus in subsequent runs we do not force it. However the Eclipse preferences for Automatic Update are saved in (so called) Profile scope. So if the user deletes and reinstalls the Aptana Studio 3 and opens the same work space (which is likely) the Automatic Update remains turned off from then on.